Updated server to host webpage
Had to change streamreader and writer to be local variables.

New endpoints to work with webpage calls.

The JPG image doesn't want to be transfered due to memory issues.
Perhaps it is possible to increase some memory block or allocation
to make it happen or use a smaller sized image.